core-connection: split variant-dependent confinement test into source sets (#1234 item 6 round 3)
The 'default guard follows BuildConfig.DEBUG' case hard-asserted DEBUG in the
common src/test suite, which the CI Unit job also compiles into testReleaseUnitTest
where BuildConfig.DEBUG is false -> release-variant failure. Split it: the
armed-default case moves to src/testDebug, the no-op-default case to src/testRelease,
so each variant asserts its real contract with no assumeTrue skip and no dropped
coverage. Test-only; production guard unchanged.
